==> 0. Querying the Twitter API.ipynb <==
Import the necessary packages

import numpy as np
import pandas as pd
import tweepy
import pprint, time
from math import ceil

Import the original csv file from the GitHub account, which is in this folder.

hs = pd.read_csv('hatespeech-master/NAACL_SRW_2016.csv', index_col=0, header=None)
hs = hs.reset_index()

Rename column headings, reset ID to not be the ID number

hs = hs.rename(columns={0: 'ID', 1: 'Class'})
hs.Class = hs.Class.astype('str')
#print(hs.head())

Determine how big the dataset is (rows, columns)

len(hs)

Query the Twitter API to get the tweets from the IDS. Get various key's from Twitter developer website (dev.twitter.com). Run 167 loops of 100 tweets each: first, subset the dataframe by the 100 tweets of each iteration (line 14). Then, define an object "tweets" which actually queries the Twitter API for the IDs (fed in list form) of that subset of tweets. Then add those tweets to an existing list, tweet_list, initialized to an empty list in line 10. Finally, sleep for 60 seconds between each loop to avoid breaking the API rate limit. 

####CHANGE THE KEYS TO YOUR OWN USING THE TWITTER API####

consumer_key = "Use Yours"
consumer_secret = "Use Yours"
access_token = "Use Yours"
access_token_secret = "Use Yours"
auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_token, access_token_secret)

api = tweepy.API(auth)

tweet_list = []


for i in range(0, ceil(len(hs.ID) / 100)):
    ids = hs.ID.iloc[i*100 : (i+1)*100]
    tweets = api.statuses_lookup(list(ids), True)
    tweet_list.extend(list(tweets))
    time.sleep(60)

Choose one tweet and print an object from that tweet for testing purposes, here the first tweet t and its object "text." 

t = tweet_list[0]
print(t.text)

After you've run the querying code for at least three hours, you can create lists of various objects from the tweets. In this function we add all of the attributes we were interested in for this project. To see all of the possible attributes, check out the Twitter API's website.

def get_tweet_info(tweet_list):
    len_list = len(tweet_list)
    tweet = []
    id_str = []
    favorite_count = []
    created_at = []
    screen_name = []
    location = []
    description = []
    followers_count = []
    hashtags = []
    user_mentions = []
    for t in tweet_list:
        tweet.append(t.text)
        id_str.append(t.id_str)
        favorite_count.append(t.favorite_count)
        created_at.append(t.created_at)
        screen_name.append(t.author.screen_name)
        location.append(t.author.location)
        description.append(t.author.description)
        followers_count.append(t.author.followers_count)
        hashtags.append(t.entities['hashtags'])
        user_mentions.append(t.entities['user_mentions'])
    tweet_info = pd.DataFrame({'Tweets': tweet, "ID": id_str, 
                               "Authors": screen_name, "Locations": location, 
                               "Descriptions": description,"Follower Count":followers_count, 
                               "Time Tweeted": created_at, "User Mentions": user_mentions, 
                               "Hashtags": hashtags, "Favorite Count":favorite_count})
    return tweet_info

df = get_tweet_info(tweet_list)
#df.head

And authors. Here I've also included code for how to write the list to a csv file so that we can use it later.

df.to_csv('hatespeech.csv')
