==> 2. TF-IDF.ipynb <==
Import necessary packages.

import nltk
import re
import string
import pandas as pd

Read in the csvs we created and convert them into lists. 

sexism = pd.read_csv('sexist_tokens.csv', encoding = "ISO-8859-1", index_col=0)
racism = pd.read_csv('racist_tokens.csv', encoding = "ISO-8859-1", index_col=0)
other = pd.read_csv('other_tokens.csv', encoding = "ISO-8859-1", index_col=0)

sexism = list(sexism['Sexist tokens'])
racism = list(racism['Racist tokens'])
other = list(other['Other'])

Convert the term vectors into a gensim dictionary, then create the TFIDF vectors from the bag-of-words corpora

#  Convert term vectors into gensim dictionary

dict = gensim.corpora.Dictionary(term_vec)

corp = []
for i in range(0, len(term_vec)):
    corp.append( dict.doc2bow( term_vec[ i ] ) )

#  Create TFIDF vectors based on term vectors bag-of-word corpora

tfidf_model = gensim.models.TfidfModel( corp )

tfidf = [ ]
for i in range( 0, len( corp ) ):
    tfidf.append( tfidf_model[ corp[ i ] ] )

dictionary = dict.token2id

Create a list called "terms." this is where we'll be appending the different TFIDF words for each list. (Terms will be a list of lists.) This for loop prints the words with the top 100 TFIDF values for each list, by sorting them and then taking the first 100 of the list, then matching the indices to the gensim dictionary above. We append each term into an internal list, then append the whole list into the terms list of lists.

terms = []

for i in range(0,len(tfidf)):
    term = []
    sorted_values = sorted(tfidf[i], key=lambda x: x[1])
    first = (len(sorted_values))-99
    top_100 = sorted_values[first:]
    values =  [int(i[0]) for i in top_100]
    num = i+1
    print("Doc",str(num),"terms of note:")
    for key,val in dictionary.items():
        if val in values:
            print(key)
            term.append(key)
    print()
    terms.append(term)

Create a dataframe of tfidf terms and save that dataframe as a csv file to access in the next Jupyter notebook.

tfidf = pd.DataFrame({'Sexist_terms': terms[0], 'Racist_terms': terms[1],'Other_terms': terms[2]})
tfidf.to_csv('tfidf.csv')
