==> 4. Neural Network.ipynb <==
Import necessary packages.

import numpy as np
import pandas as pd
from keras.layers import Dense
from keras.models import Sequential
from keras.callbacks import EarlyStopping
from keras.utils import to_categorical

Read in the final dataset, with all of our variables.

df = pd.read_csv('full_dataset.csv', encoding = "ISO-8859-1", index_col=0)

For this version of the neural net, we decided to keep only the binary variables for tfidf and the hashtag variables, in addition to the negative sentiment and compound sentiment variables as predictors. These are variables 11 through 1576. Create a dataframe out of only these columns below. To get an idea of what this looks like, uncomment the print statement.

rel_var = df.iloc[:,11:1576]
print(rel_var.head())

Because keras requires a numpy array as input, we convert the dataframe of our predictors to a numpy array below. To see what this looks like, uncomment the print statement.

predictors = pd.DataFrame.as_matrix(rel_var, columns=None)
#print(predictors)

Here, we change our class variables to numeric values for the keras package to predict. To compare the lists, uncomment the print statement. We then use keras's to_categorical function on the numeric values to show that we're predicting multiple classes. n_cols is the number columns in the predictors dataframe, which will be fed as an argument to the model.add() function below

new_class = []

for i in df.Class:
    if i == 'sexism':
        new_class.append(0)
    elif i == 'racism':
        new_class.append(1)
    else:
        new_class.append(2)

#print(list(df.Class)[0:10])
#print(new_class[0:10])

target=to_categorical(new_class)
n_cols = predictors.shape[1]

Here we define the architecture of the model. Our model is a sequential model, and each model.add() line represents a layer in the neural network; the number after Dense( represents the number of nodes in that layer. Our activation function is "relu" for each layer, and "softmax" for the last layer (because it's predicting a categorical variable, this tells it to predict whichever class has the highest probability of being true). 

#Create model
model = Sequential()

#Add layers to model
model.add(Dense(100, activation='relu', input_shape = (n_cols,)))
model.add(Dense(1000, activation='relu'))
model.add(Dense(1000, activation='relu'))
model.add(Dense(1000, activation='relu'))
model.add(Dense(1000, activation='relu'))
model.add(Dense(1000, activation='relu'))
model.add(Dense(1000, activation='relu'))
model.add(Dense(3, activation='softmax'))

Finally, we run and compile the code. We use the adam optimizer, a loss function of categorical crossentropy (used for categorical target variables), and our metric of interest is accuracy. We also add an early stopping function with a patience of two, meaning that if our model fails to improve on our metric for two epochs, the model should stop running. The model fit function is the final piece of important code. It takes as input our predictors, our target variable, a validation split (what proportion of the dataset should be set aside for testing), and the number of epochs (this can be higher since we have an early stopping function). The accuracy of the model is the highest number under "val_acc." 

#compile 
model.compile(optimizer='adam', loss='categorical_crossentropy', metrics = ['accuracy'])

early_stop = EarlyStopping(patience=2) 

#fit model
model.fit(predictors, target, validation_split=0.3, epochs=20, callbacks=[early_stop])
